refactor!: change lexer to use acorn and escodegen
- refactor!: change lexer to use acorn and codegen instead of js-tokens.
Realized that creating our own AST reader is a huge task, make take
a lot of time and prone for errors if we want to be spec compliant.
With this, I decided to use acorn and codegen instead in terms of
identifying variables.
- fix: non-MiniJS variables are being read by Mini.js.
I made so that we are not automatically reading user
defined global variables. We now only fetch variables from
attributes and events. Currently, we can't fetch from
the scripts since we are not sure whether that script is
defined by a user or a library.
- fix: not being able to declare variables due to being read
as a proxy variable.